Move Back Office Overview page utils to separate files as well
