+++ apps/back-office/src/app/_components/OverviewDesktop.tsx
@@ -4,7 +4,8 @@ import { Table } from '@meldingen/ui'
 
 import type { MeldingWithAddress } from '../Overview'
 
-import { getOverviewFieldLabel, OVERVIEW_FIELDS, renderOverviewFieldValue } from './utils/overviewFields'
+import { OVERVIEW_FIELDS } from '../constants'
+import { getOverviewFieldLabel, renderOverviewFieldValue } from './_utils'
 
 import styles from './OverviewDesktop.module.css'
