+++ apps/back-office/src/app/_components/_utils/getOverviewFieldLabel.ts
@@ -1,0 +1,6 @@
+import type { OVERVIEW_FIELDS } from '../../constants'
+
+export type OverviewField = (typeof OVERVIEW_FIELDS)[number]
+
+export const getOverviewFieldLabel = (field: OverviewField, t: (key: string) => string) =>
+  t(`overview.${field.labelKey}`)
