+++ apps/back-office/src/app/constants.ts
@@ -1,0 +1,9 @@
+export const OVERVIEW_FIELDS = [
+  { key: 'public_id', labelKey: 'column-header.public_id' },
+  { key: 'created_at', labelKey: 'column-header.created_at' },
+  { key: 'classification', labelKey: 'column-header.classification' },
+  { key: 'state', labelKey: 'column-header.state' },
+  { key: 'urgency', labelKey: 'column-header.urgency' },
+  { key: 'address', labelKey: 'column-header.address' },
+  { key: 'postal_code', labelKey: 'column-header.postal_code' },
+] as const
